requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
boto3==1.16.18
botocore==1.19.18
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gdown==3.12.2
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
libcst==0.3.13
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
nltk==3.5
notebook==6.1.5
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
proto-plus==1.11.0
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
sklearn==0.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.9.3
torch==1.7.0+cu110
torchaudio==0.7.0
torchvision==0.8.1+cu110
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
transformers==3.5.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
